requirements.txt
streamlit
openai
python-dotenv
qrcode
pillow
firebase-admin